+++ http/payload/requests.go
@@ -35,7 +35,7 @@ type LoginRequest struct {
 }
 
 type LoginResponse struct {
-	Id          int64  `json:"id"`
+	Id          uint   `json:"id"`
 	Token       string `json:"token"`
 	ApiKey      string `json:"apiKey,omitempty"`
 	Permissions int    `json:"permissions"`
